@@ -17,42 +17,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.0.1",
-    "@angular/common": "7.0.1",
-    "@angular/compiler": "7.0.1",
-    "@angular/core": "7.0.1",
-    "@angular/forms": "7.0.1",
-    "@angular/http": "7.0.1",
-    "@angular/platform-browser": "7.0.1",
-    "@angular/platform-browser-dynamic": "7.0.1",
-    "@angular/router": "7.0.1",
-    "@ngrx/effects": "6.1.0",
-    "@ngrx/router-store": "6.1.0",
-    "@ngrx/store": "6.1.0",
-    "@ngrx/store-devtools": "6.1.0",
-    "@practicalwebdev/utils": "2.0.1",
+    "@angular/animations": "7.2.15",
+    "@angular/common": "7.2.15",
+    "@angular/compiler": "7.2.15",
+    "@angular/core": "7.2.15",
+    "@angular/forms": "7.2.15",
+    "@angular/http": "7.2.15",
+    "@angular/platform-browser": "7.2.15",
+    "@angular/platform-browser-dynamic": "7.2.15",
+    "@angular/router": "7.2.15",
+    "@lernato/common": "1.2.0",
+    "@lernato/common-angular": "1.1.0",
+    "@ngrx/effects": "7.4.0",
+    "@ngrx/router-store": "7.4.0",
+    "@ngrx/store": "7.4.0",
+    "@ngrx/store-devtools": "7.4.0",
     "core-js": "2.5.7",
-    "rxjs": "6.3.3",
-    "zone.js": "0.8.26"
+    "rxjs": "6.5.2",
+    "tslib": "1.9.0",
+    "zone.js": "0.8.29"
   },
   "devDependencies": {
     "@angular-builders/jest": "1.3.1",
-    "@angular-devkit/build-angular": "0.10.3",
-    "@angular/cli": "7.0.3",
-    "@angular/compiler-cli": "7.0.1",
-    "@angular/language-service": "7.0.1",
-    "@practicalwebdev/schematics": "2.0.1",
+    "@angular-devkit/build-angular": "0.13.0",
+    "@angular/cli": "7.3.9",
+    "@angular/compiler-cli": "7.2.15",
+    "@angular/language-service": "7.2.15",
+    "@lernato/schematics": "1.2.0",
     "@types/jasmine": "2.8.9",
     "@types/jasminewd2": "2.0.5",
     "@types/node": "10.12.0",
     "codelyzer": "4.5.0",
     "jasmine-spec-reporter": "4.2.1",
     "jest": "23.6.0",
-    "jest-sonar-reporter": "2.0.0",
     "lite-server": "2.4.0",
     "protractor": "5.4.1",
     "ts-node": "7.0.1",
     "tslint": "5.11.0",
-    "typescript": "3.1.3"
+    "typescript": "3.2.4"
   }
 }
