@@ -1,20 +1,15 @@
 import { Injectable } from '@angular/core';
+import { AppState, Route, StoreService } from '@lernato/common-angular';
 import { RouterReducerState } from '@ngrx/router-store';
 import { createFeatureSelector, createSelector, Store } from '@ngrx/store';
+import { BackAction, ForwardAction, GoAction, RouteChangeAction } from './router.store.actions';
 
-import { AppState, Route, StoreService } from '@practicalwebdev/utils';
 
-import {
-  BackAction,
-  ForwardAction,
-  GoAction,
-  RouteChangeAction
-} from './router.store.actions';
 
 @Injectable()
 export class RouterStoreService extends StoreService {
   routerReducerState = createFeatureSelector<RouterReducerState>('router');
-  routerPathSelector = createSelector(this.routerReducerState, this.getProperty('url'));
+  routerPathSelector = createSelector(this.routerReducerState, state => state['url']);
 
   constructor(public store: Store<AppState>) { super(); }
 
