@@ -1,15 +1,15 @@
 import { Injectable } from '@angular/core';
+import { StoreService } from '@lernato/common-angular';
 import { createFeatureSelector, createSelector, Store } from '@ngrx/store';
-import { StoreService } from '@practicalwebdev/utils';
-
 import { VillainStoreState } from '../../models';
-
 import { LoadAction, LoadFailAction, LoadSuccessAction } from './villain.store.actions';
+
+
 
 @Injectable()
 export class VillainStoreService extends StoreService {
   villainStoreState = createFeatureSelector<VillainStoreState>('villains');
-  villainsSelector = createSelector(this.villainStoreState, this.getProperty('villains'));
+  villainsSelector = createSelector(this.villainStoreState, state => state.villains);
 
   constructor(public store: Store<VillainStoreState>) { super(); }
 
